requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
factory-boy==2.11.1
Faker==1.0.2
Flask==1.0.2
Flask-Caching==1.4.0
Flask-Classful==0.14.2
Flask-Inputs==0.3.0
Flask-Login==0.4.1
flask-marshmallow==0.9.0
Flask-SQLAlchemy==2.3.2
future==0.17.1
httpie==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
marshmallow==2.18.0
marshmallow-jsonschema==0.5.0
marshmallow-sqlalchemy==0.15.0
more-itertools==5.0.0
pluggy==0.8.1
psycopg2-binary==2.7.7
py==1.7.0
Pygments==2.3.1
pytest==4.2.0
pytest-flask==0.14.0
python-dateutil==2.7.5
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.17
text-unidecode==1.2
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
